requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
google-genai>=1.31.0
streamlit>=1.32.0
pandas>=2.1.0